Add 3 new types of task (Todo, Deadline, Event)
Used inheritance to extend the class "Task" into 3 new types of tasks. Added specific variables that each type of task requires in the new classes. Used polymorphism and method overloading so that when the "TaskManager" class requests for getTaskType(), the appropriate task type is returned.

Improved parser so that it methodically splits the text input into the (Input Command, context of text, appropriate deadlines/from/to), allowing for better seperation into different variables (parsing method can be more efficient though)